Extract util function into its own module.
