+++ cmd/soroban-cli/src/commands/contract/deploy.rs
@@ -1,6 +1,7 @@
 use crate::commands::global;
 
 pub mod asset;
+pub mod utils;
 pub mod wasm;
 
 #[derive(Debug, clap::Subcommand)]
