+++ cmd/soroban-cli/src/commands/contract/deploy/asset.rs
@@ -21,6 +21,8 @@ use crate::{
     utils::contract_id_hash_from_asset,
 };
 
+use crate::commands::contract::deploy::utils::alias_validator;
+
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
     #[error("error parsing int: {0}")]
@@ -56,8 +58,15 @@ pub struct Cmd {
 
     #[command(flatten)]
     pub config: config::Args,
+
     #[command(flatten)]
     pub fee: crate::fee::Args,
+
+    /// The alias that will be used to save the assets's id.
+    /// Whenever used, `--alias` will always overwrite the existing contract id
+    /// configuration without asking for confirmation.
+    #[arg(long, value_parser = clap::builder::ValueParser::new(alias_validator))]
+    pub alias: Option<String>,
 }
 
 impl Cmd {
