+++ cmd/soroban-cli/src/commands/contract/deploy/wasm.rs
@@ -1,3 +1,4 @@
+use crate::commands::contract::deploy::utils::alias_validator;
 use std::array::TryFromSliceError;
 use std::ffi::OsString;
 use std::fmt::Debug;
@@ -12,7 +13,6 @@ use crate::xdr::{
 };
 use clap::{arg, command, Parser};
 use rand::Rng;
-use regex::Regex;
 
 use soroban_spec_tools::contract as contract_spec;
 
@@ -149,18 +149,6 @@ impl Cmd {
             }
         }
         Ok(())
-    }
-}
-
-fn alias_validator(alias: &str) -> Result<String, Error> {
-    let regex = Regex::new(r"^[a-zA-Z0-9_-]{1,30}$").unwrap();
-
-    if regex.is_match(alias) {
-        Ok(alias.into())
-    } else {
-        Err(Error::InvalidAliasFormat {
-            alias: alias.into(),
-        })
     }
 }
 
@@ -389,35 +377,5 @@ mod tests {
         );
 
         assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_alias_validator_with_valid_inputs() {
-        let valid_inputs = [
-            "hello",
-            "123",
-            "hello123",
-            "hello_123",
-            "123_hello",
-            "123-hello",
-            "hello-123",
-            "HeLlo-123",
-        ];
-
-        for input in valid_inputs {
-            let result = alias_validator(input);
-            assert!(result.is_ok());
-            assert!(result.unwrap() == input);
-        }
-    }
-
-    #[test]
-    fn test_alias_validator_with_invalid_inputs() {
-        let invalid_inputs = ["", "invalid!", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"];
-
-        for input in invalid_inputs {
-            let result = alias_validator(input);
-            assert!(result.is_err());
-        }
     }
 }
